requirements.txt
streamlit
csv
numpy
pandas
PyMySQL-1.0.2